+++ apps/meteor/client/providers/AppsProvider.tsx
@@ -1,5 +1,5 @@
 import { useDebouncedCallback } from '@rocket.chat/fuselage-hooks';
-import { usePermission, useSingleStream } from '@rocket.chat/ui-contexts';
+import { usePermission, useStream } from '@rocket.chat/ui-contexts';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import type { FC } from 'react';
 import React, { useEffect } from 'react';
@@ -8,11 +8,25 @@ import { AppClientOrchestratorInstance } from '../../ee/client/apps/orchestrator
 import { AppsContext } from '../contexts/AppsContext';
 import { useIsEnterprise } from '../hooks/useIsEnterprise';
 import { useInvalidateLicense } from '../hooks/useLicense';
+import type { AsyncState } from '../lib/asyncState';
 import { AsyncStatePhase } from '../lib/asyncState';
 import { useInvalidateAppsCountQueryCallback } from '../views/marketplace/hooks/useAppsCountQuery';
 import type { App } from '../views/marketplace/types';
 
 const sortByName = (apps: App[]): App[] => apps.sort((a, b) => (a.name.toLowerCase() > b.name.toLowerCase() ? 1 : -1));
+
+const getAppState = (
+	loading: boolean,
+	apps: App[] | undefined,
+): Omit<
+	AsyncState<{
+		apps: App[];
+	}>,
+	'error'
+> => ({
+	phase: loading ? AsyncStatePhase.LOADING : AsyncStatePhase.RESOLVED,
+	value: { apps: apps || [] },
+});
 
 const AppsProvider: FC = ({ children }) => {
 	const isAdminUser = usePermission('manage-apps');
@@ -25,7 +39,7 @@ const AppsProvider: FC = ({ children }) => {
 	const invalidateAppsCountQuery = useInvalidateAppsCountQueryCallback();
 	const invalidateLicenseQuery = useInvalidateLicense();
 
-	const stream = useSingleStream('apps');
+	const stream = useStream('apps');
 
 	const invalidate = useDebouncedCallback(
 		() => {
@@ -109,7 +123,11 @@ const AppsProvider: FC = ({ children }) => {
 				};
 
 				if (installedApp) {
-					installedApps.push(record);
+					if (installedApp.private) {
+						privateApps.push(record);
+					} else {
+						installedApps.push(record);
+					}
 				}
 
 				marketplaceApps.push(record);
@@ -129,13 +147,16 @@ const AppsProvider: FC = ({ children }) => {
 		},
 	);
 
+	const [marketplaceAppsData, installedAppsData, privateAppsData] = store.data || [];
+	const { isLoading } = store;
+
 	return (
 		<AppsContext.Provider
 			children={children}
 			value={{
-				installedApps: { phase: AsyncStatePhase.RESOLVED, value: { apps: store.data?.[1] || [] } },
-				marketplaceApps: { phase: AsyncStatePhase.RESOLVED, value: { apps: store.data?.[0] || [] } },
-				privateApps: { phase: AsyncStatePhase.RESOLVED, value: { apps: store.data?.[2] || [] } },
+				installedApps: getAppState(isLoading, installedAppsData),
+				marketplaceApps: getAppState(isLoading, marketplaceAppsData),
+				privateApps: getAppState(isLoading, privateAppsData),
 				reload: async () => {
 					await Promise.all([queryClient.invalidateQueries(['marketplace'])]);
 				},