+++ apps/meteor/client/views/admin/settings/inputs/ActionSettingInput.stories.tsx
@@ -1,9 +1,8 @@
 import { Field } from '@rocket.chat/fuselage';
-import type { ServerMethods } from '@rocket.chat/ui-contexts';
+import type { ServerMethods, TranslationKey } from '@rocket.chat/ui-contexts';
 import type { ComponentMeta, ComponentStory } from '@storybook/react';
 import React from 'react';
 
-import type keys from '../../../../../packages/rocketchat-i18n/i18n/en.i18n.json';
 import ActionSettingInput from './ActionSettingInput';
 
 export default {
@@ -17,14 +16,14 @@ const Template: ComponentStory<typeof ActionSettingInput> = (args) => <ActionSet
 export const Default = Template.bind({});
 Default.args = {
 	_id: 'setting_id',
-	actionText: 'Action text' as keyof typeof keys,
+	actionText: 'Action text' as TranslationKey,
 	value: 'methodName' as keyof ServerMethods,
 };
 
 export const Disabled = Template.bind({});
 Disabled.args = {
 	_id: 'setting_id',
-	actionText: 'Action text' as keyof typeof keys,
+	actionText: 'Action text' as TranslationKey,
 	value: 'methodName' as keyof ServerMethods,
 	disabled: true,
 };
@@ -32,7 +31,7 @@ Disabled.args = {
 export const WithinChangedSection = Template.bind({});
 WithinChangedSection.args = {
 	_id: 'setting_id',
-	actionText: 'Action text' as keyof typeof keys,
+	actionText: 'Action text' as TranslationKey,
 	value: 'methodName' as keyof ServerMethods,
 	sectionChanged: true,
 };