+++ apps/meteor/ee/client/omnichannel/reports/components/index.ts
@@ -2,4 +2,6 @@ export * from './AgentsTable';
 export * from './BarChart';
 export * from './PieChart';
 export * from './ReportCard';
-export * from './CardErrorState';
+export * from './ReportCardErrorState';
+export * from './ReportCardLoadingState';
+export * from './ReportCardEmptyState';