+++ apps/meteor/tests/e2e/page-objects/fragments/home-content.ts
@@ -149,7 +149,7 @@ export class HomeContent {
 	}
 
 	get btnRecordAudio(): Locator {
-		return this.page.locator('[data-qa-id="audio-record"]');
+		return this.page.locator('[data-qa-id="audio-message"]');
 	}
 
 	get btnMenuMoreActions() {
@@ -196,14 +196,6 @@ export class HomeContent {
 		return this.page.locator('.rcx-vertical-bar button:has-text("Create")');
 	}
 
-	get inputModalAgentUserName(): Locator {
-		return this.page.locator('#modal-root input[placeholder="Username, name or e-mail"]');
-	}
-
-	get inputModalAgentForwardComment(): Locator {
-		return this.page.locator('[data-qa-id="ForwardChatModalTextAreaInputComment"]');
-	}
-
 	async pickEmoji(emoji: string, section = 'Smileys & People') {
 		await this.page.locator('role=toolbar[name="Composer Primary Actions"] >> role=button[name="Emoji"]').click();
 
@@ -245,10 +237,6 @@ export class HomeContent {
 		await this.page.locator('//main//aside >> [name="alsoSendThreadToChannel"]').setChecked(isChecked);
 	}
 
-	get takeOmnichannelChatButton(): Locator {
-		return this.page.locator('role=button[name="Take it!"]');
-	}
-
 	get lastSystemMessageBody(): Locator {
 		return this.page.locator('[data-qa-type="system-message-body"]').last();
 	}
@@ -287,5 +275,9 @@ export class HomeContent {
 
 	get btnAnonymousTalk(): Locator {
 		return this.page.locator('role=button[name="Or talk as anonymous"]');
+	}
+
+	findSystemMessage(text: string): Locator {
+		return this.page.locator(`[data-qa-type="system-message-body"] >> text="${text}"`);
 	}
 }