+++ ee/packages/license/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @rocket.chat/license
 
+## 0.1.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.3
+
+## 0.1.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.2
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [c2b224fd82]
+  - @rocket.chat/core-typings@6.5.1
+
 ## 0.1.0
 
 ### Minor Changes