+++ packages/core-services/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @rocket.chat/core-services
 
+## 0.3.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.3
+- @rocket.chat/rest-typings@6.5.3
+- @rocket.chat/models@0.0.27
+
+## 0.3.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.2
+- @rocket.chat/rest-typings@6.5.2
+- @rocket.chat/models@0.0.26
+
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [c2b224fd82]
+- Updated dependencies [c2b224fd82]
+  - @rocket.chat/rest-typings@6.5.1
+  - @rocket.chat/core-typings@6.5.1
+  - @rocket.chat/models@0.0.25
+
 ## 0.3.0
 
 ### Minor Changes