+++ packages/ui-contexts/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @rocket.chat/ui-contexts
 
+## 3.0.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.3
+- @rocket.chat/rest-typings@6.5.3
+- @rocket.chat/ddp-client@0.2.12
+
+## 3.0.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.2
+- @rocket.chat/rest-typings@6.5.2
+- @rocket.chat/ddp-client@0.2.11
+
+## 3.0.1
+
+### Patch Changes
+
+- Updated dependencies [c2b224fd82]
+- Updated dependencies [c2b224fd82]
+- Updated dependencies [c2b224fd82]
+- Updated dependencies [c2b224fd82]
+- Updated dependencies [c2b224fd82]
+- Updated dependencies [c2b224fd82]
+- Updated dependencies [c2b224fd82]
+  - @rocket.chat/rest-typings@6.5.1
+  - @rocket.chat/ddp-client@0.2.10
+  - @rocket.chat/core-typings@6.5.1
+
 ## 3.0.0
 
 ### Patch Changes