+++ packages/web-ui-registration/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @rocket.chat/web-ui-registration
 
+## 3.0.3
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.3
+
+## 3.0.2
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.1
+
 ## 3.0.0
 
 ### Patch Changes